Refactor secure area key options and purpose
This commit refactors the secure area key options and purpose. It introduces a new file `KeyOptions.swift` that defines the tasks for which keys can be used, such as signing and key agreement. It also adds a new enum `KeyAccessProtection` to specify the condition for the key to be available. Additionally, the `KeyPurpose.swift` file has been deleted as it is no longer needed.
